requirements.txt
fastapi
uvicorn
pydantic
asyncpg
aiohttp
beautifulsoup4
feedparser
requests
asyncio
python-dateutil
trafilatura
sqlalchemy
httpx
environs
pyyaml
pytest
pandas
numpy
matplotlib
seaborn
jinja2
python-multipart
starlette
scrapy
tiktoken
transformers
tenacity
python-dotenv
colorlog
tqdm
ujson
colorama
pytest-asyncio
ujson